@@ -2,55 +2,50 @@
   "name": "apis",
   "version": "1.0.0",
   "description": "",
-  "main": "./src/server.ts",
+  "main": "./src/app.ts",
   "scripts": {
-    "lint": "eslint",
+    "lint": "eslint \".\\src\\**\"",
     "tsc": "tsc",
-    "dev": "ts-node-dev --respawn --transpileOnly ./src/server.ts",
-    "start": "npm run src/server.js",
-    "prod": "tsc && copy package.json .\\build\\ && copy  package-lock.json .\\build\\ && copy config.json .\\build\\ && node ./build/src/server.js"
+    "dev": "ts-node-dev --respawn --transpileOnly ./src/app.ts",
+    "start": "npm run src/app.js",
+    "prod": "tsc && copy package.json .\\build\\ && copy  package-lock.json .\\build\\ && copy config.json .\\build\\ && node ./build/src/app.js"
   },
   "author": "",
   "license": "ASE IT POC",
+  "type": "commonjs",
+  "engines": {
+    "node": ">=12.18.0"
+  },
   "dependencies": {
-    "@types/express": "^4.16.1",
-    "@types/node": "7.0.22",
-    "@types/swagger-ui-express": "^3.0.0",
     "body-parser": "^1.19.0",
     "cors": "^2.8.5",
-    "express": "^4.16.4",
-    "express-openapi-validate": "^0.4.4",
-    "fs": "0.0.1-security",
-    "js-yaml": "^3.13.1",
-    "json-loader": "^0.5.7",
+    "express": "^4.17.1",
+    "express-openapi-validate": "^0.5.1",
+    "got": "^11.3.0",
     "jsonwebtoken": "^8.5.1",
-    "mongodb": "^2.2.33",
-    "nodemon": "^1.19.2",
-    "path": "^0.12.7",
-    "querystring": "^0.2.0",
-    "request": "^2.88.0",
-    "request-promise": "^4.2.4",
+    "mongodb": "^3.5.8",
     "sessionstorage": "^0.1.0",
-    "swagger-jsdoc": "^3.2.9",
-    "swagger-ui-express": "^4.0.4",
-    "ts-node-dev": "^1.0.0-pre.30",
-    "uuid": "^3.4.0"
+    "swagger-jsdoc": "^4.0.0",
+    "swagger-ui-express": "^4.1.4",
+    "uuid": "^8.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.5",
-    "@babel/preset-env": "^7.4.5",
-    "babel-loader": "^8.0.6",
-    "ts-loader": "^6.0.1",
-    "typescript": "^3.3.3333",
-    "webpack": "^4.32.2",
-    "webpack-cli": "^3.3.2",
-    "webpack-dev-server": "^3.4.1"
-  },
-  "browser": {
-    "colors": false,
-    "events": "eventemitter3",
-    "checkForModuleDuplicates": false,
-    "url": "./shims/url",
-    "tracer": false
+    "@babel/core": "^7.10.2",
+    "@babel/preset-env": "^7.10.2",
+    "@types/cors": "^2.8.6",
+    "@types/express": "^4.17.6",
+    "@types/jsonwebtoken": "^8.5.0",
+    "@types/mongodb": "^3.5.20",
+    "@types/node": "^12.0.0",
+    "@types/swagger-ui-express": "^4.1.2",
+    "@types/uuid": "^8.0.0",
+    "@typescript-eslint/eslint-plugin": "^3.1.0",
+    "@typescript-eslint/parser": "^3.1.0",
+    "babel-loader": "^8.1.0",
+    "eslint": "^7.1.0",
+    "eslint-plugin-node": "^11.1.0",
+    "ts-loader": "^7.0.5",
+    "ts-node-dev": "^1.0.0-pre.44",
+    "typescript": "^3.9.3"
   }
 }
