@@ -1,7 +1,7 @@
 import { ActivePort } from '../class/ActivePort';
 import { BaseLayer } from './BaseLayer';
-const httppromise = require('request-promise');
-var sessionstorage = require('sessionstorage');
+import httppromise, { Options } from 'got';
+import sessionstorage from 'sessionstorage';
 import { Log } from '../class/Log'
 import { Logger } from '../class/Logger'
 import { EnumCurrentStatus } from '../Enum/EnumCurrentStatus'
@@ -15,55 +15,53 @@ export class ActivePortBaseLayer extends BaseLayer {
         super();
     }
 
-    baseUrl(url: string) {
+    baseUrl(url: string): string {
         return this.environmentConfig.ActivePort.Urls.BaseUrl + url;
     }
 
     protected activePort: ActivePort = new ActivePort();
 
     //Generate new Token for ActivePort
-    private generateActivePortToken() {
+    private async generateActivePortToken() {
 
         this.activePort.Username = this.environmentConfig.ActivePort.Username;
         this.activePort.Password = this.environmentConfig.ActivePort.Password;
-        
-        let body = {
+
+        const body = {
             "username": this.activePort.Username,
             "password": this.activePort.Password,
             "rememberMe": true
         };
 
-        let options = {
+        const options: Options = {
             url: this.baseUrl(Constants.ActivePortAuthURL),
             method: 'POST',
-            json: true,
-            body: body
+            json: body
         };
-        return httppromise(options);
-    };
+        return await httppromise(options);
+    }
 
 
     //Check if ActivePort token already exists otherwise it will generate new Token for ActivePort
-    protected authorizeActivePort() {
+    protected async authorizeActivePort(): Promise<boolean> {
         if (sessionstorage.getItem(EnumToken.ActivePortToken) == null) {
-            
-                return this.generateActivePortToken().then(function (response: any) {
+            try {
+                const response: any = await this.generateActivePortToken();
                 sessionstorage.setItem(EnumToken.ActivePortToken, response.id_token)
                 Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.ActivePort, Constants.ActivePortAuthSuccess, response, ""))
                 return true;
-            }).catch(function (err: any) {
-                console.log(err)
-
+            } catch (err) {
+                console.error(err)
                 Logger.updateLogs(new Log(EnumCurrentStatus.Error, EnumModule.ActivePort, Constants.ActivePortAuthError, err, ""));
                 return false;
-            })
-            //return authorizeResponse;
+            }
         }
+        return false;
     }
 
 
     //isActivePortAuthorized token
-    protected async isActivePortAuthorized() {
+    protected async isActivePortAuthorized(): Promise<boolean> {
         if (await Logger.hasErrorLogs() == true)
             return false;
         else if (await this.authorizeActivePort() == false)
@@ -73,12 +71,11 @@ export class ActivePortBaseLayer extends BaseLayer {
     }
 
     //remove token
-    protected removeToken() {
-      try {
+    protected removeToken(): void {
+        try {
             sessionstorage.removeItem(EnumToken.ActivePortToken);
             Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.ActivePort, Constants.ActivePortTokenRemovedSuccess, "", ""))
-        }
-        catch (error) {
+        } catch (error) {
             console.error(error);
             Logger.updateLogs(new Log(EnumCurrentStatus.Error, EnumModule.ActivePort, Constants.ActivePortTokenRemovedError, "", ""));
         }
