@@ -1,17 +1,14 @@
 import { BaseLayer } from "./BaseLayer";
-
-var jwt = require('jsonwebtoken');
-const fs = require('fs');
-var sessionstorage = require('sessionstorage');
+import jwt from 'jsonwebtoken';
 
 export class Authentication extends BaseLayer {
 
-    async Authorize(username: string, password: string) {
+    public Authorize(username: string, password: string): string {
 
         if (username == "munish" && password == "singla") {
-            let payload = { username: username, password: password };
+            const payload = { username, password };
 
-            var signOptions = {
+            const signOptions = {
                 issuer: this.environmentConfig.PrivateKey.Issuer,
                 subject: this.environmentConfig.PrivateKey.Subject,
                 audience: this.environmentConfig.PrivateKey.Audience,
@@ -19,25 +16,20 @@ export class Authentication extends BaseLayer {
                 //algorithm: "RS256"   // RSASSA [ "RS256", "RS384", "RS512" ]
             };
 
-            var token = jwt.sign(payload, this.environmentConfig.PrivateKey.Value, signOptions);
-            //sessionstorage.setItem("Authorization", token);
-            //console.log("\nToken is1: " + sessionstorage.getItem("Authorization", token));
-            return token;
+            return jwt.sign(payload, this.environmentConfig.PrivateKey.Value, signOptions);
         }
         else {
             return 'invalid';
         }
     }
 
-    verifyAuthentication(token: string) {
-        //console.log("\nToken is 2: " + sessionstorage.getItem("Authorization", token));
-        var decoded = jwt.verify(token, this.environmentConfig.PrivateKey);
-        //console.log(decoded);
-        if (decoded.username == "munish" && decoded.password == "singla") {
-            return true;
+    public verifyAuthentication(token: unknown): boolean {
+        if (token && typeof token === 'string') {
+            const decoded: any = jwt.verify(token, this.environmentConfig.PrivateKey);
+            if (decoded.username == "munish" && decoded.password == "singla") {
+                return true;
+            }
         }
-        else
-            return false;
-
+        return false;
     }
 }
