@@ -1,5 +1,4 @@
-const httppromise = require('request-promise');
-var sessionstorage = require('sessionstorage');
+import httppromise, { Options } from 'got';
 import { Log } from '../class/Log'
 import { Logger } from '../class/Logger'
 import { EnumCurrentStatus } from '../Enum/EnumCurrentStatus'
@@ -21,86 +20,72 @@ class NetAppHttpRequests extends NetAppBaseLayer {
     if (await this.isNetAppAuthorized() == false)
       return;
 
-    let options = {
+    const options: Options = {
       url: this.baseUrl(Constants.NetAppClusters),
       method: 'GET',
-      headers: this.netAppHeader(),
-      json: true
+      headers: this.netAppHeader()
     };
 
-    let self = this;
-    let result;
-    await httppromise(options).then(function (response: any) {
-      result = response;
+    try {
+      const response: any = await httppromise(options);
       Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.NetApp, Constants.NetAppClusterRetrievalSuccess, response, ''));
-    }).catch(function (err: any) {
-
+      return new NetAppClustersRetrievalSuccessResponse(response);
+    } catch (err) {
       Logger.updateLogs(new Log(EnumCurrentStatus.Error, EnumModule.NetApp, Constants.NetAppClusterRetrievalFailure, err, ''));
-    })
+    }
 
-    return new NetAppClustersRetrievalSuccessResponse(result);
-  };
+  }
 
-   //Here we are creating new NKS cluster
+  //Here we are creating new NKS cluster
   //prerequisite: NetApp Token in Header
   async createNKSCluster(requestBody: any) {
 
     if (await this.isNetAppAuthorized() == false)
       return;
 
-    let body = requestBody;
+    const body = requestBody;
 
-    let options = {
+    const options: Options = {
       url: this.baseUrl(Constants.NetAppClusters),
       method: 'POST',
       headers: this.netAppHeader(),
-      body: body,
-      json: true
+      json: body
     };
 
-    let self = this;
-    let result;
-    
-    await httppromise(options).then(function (response: any) {
-      result = response;
+    try {
+      const response: any = await httppromise(options);
       Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.NetApp, Constants.NetAppClusterCreationSuccess, response, ''));
-    }).catch(function (err: any) {
-
+      return new NetAppClustersRetrievalSuccessResponse(response);
+    } catch (err) {
       Logger.updateLogs(new Log(EnumCurrentStatus.Error, EnumModule.NetApp, Constants.NetAppClusterCreationFailure, err, ''));
-    })
+    }
 
-    return new NetAppClustersRetrievalSuccessResponse(result);
-  };
+  }
 
 
-   //Here we are deleting NKS cluster
+  //Here we are deleting NKS cluster
   //prerequisite: NetApp Token in Header
   async deleteNKSCluster(param: any) {
 
     if (await this.isNetAppAuthorized() == false)
       return;
 
-    let options = {
+    const options: Options = {
       url: this.baseUrl(Constants.NetAppClusters + "/" + param.clusterid),
       method: 'DELETE',
-      headers: this.netAppHeader(),
-      json: true
+      headers: this.netAppHeader()
     };
 
-    let self = this;
-    let result;
-    
-    await httppromise(options).then(function (response: any) {
-      result = response;
+    try {
+      const response: any = await httppromise(options);
       Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.NetApp, Constants.NetAppClusterDeletionSuccess, response, ''));
-    }).catch(function (err: any) {
-
+      return new NetAppClustersRetrievalSuccessResponse(response);
+    } catch (err) {
       Logger.updateLogs(new Log(EnumCurrentStatus.Error, EnumModule.NetApp, Constants.NetAppClusterDeletionFailure, err, ''));
-    })
+    }
 
-    return new NetAppClustersRetrievalSuccessResponse(result);
-  };
+  }
+
 }
-
 
 export default new NetAppHttpRequests();
