@@ -1,5 +1,4 @@
-const httppromise = require('request-promise');
-var sessionstorage = require('sessionstorage');
+import httppromise, { Options } from 'got';
 import { Log } from '../class/Log'
 import { Logger } from '../class/Logger'
 import { EnumCurrentStatus } from '../Enum/EnumCurrentStatus'
@@ -20,15 +19,16 @@ class StitchdataHttpRequests extends StitchdataBaseLayer {
 
     this.removeToken();
     this.stitchdata.AuthorizationCode = RequestBody.stitch_auth_code;
-    let self = this;
-    await this.authorizeStitchdata(self.stitchdata).then(function (response: any) {
+
+    try {
+      const response: any = await this.authorizeStitchdata(this.stitchdata);
       Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.Stitchdata, Constants.StitchdataAuthError, response, ""));
-    }).catch(function (err: any) {
+    } catch (err) {
       Logger.updateLogs(new Log(EnumCurrentStatus.Error, EnumModule.Stitchdata, Constants.StitchdataAuthSuccess, err, ""));
-    });
+    }
 
     return new StitchdataAccountAuthorizationSuccessResponse(this.stitchdata);
-  };
+  }
 
 
   //Here we are creating tenant Account
@@ -37,7 +37,7 @@ class StitchdataHttpRequests extends StitchdataBaseLayer {
 
     this.assignStitchdataEssentials();
 
-    let body = {
+    const body = {
       "partner_id": this.stitchdata.PartnerId,
       "partner_secret": this.stitchdata.PartnerSecret,
       "first_name": RequestBody.firstname,
@@ -46,29 +46,27 @@ class StitchdataHttpRequests extends StitchdataBaseLayer {
       "email": RequestBody.email,
     }
 
-    let options = {
+    const options: Options = {
       url: this.baseUrl(Constants.StitchdataCreateAccountURL),
       method: 'POST',
       headers: {
         'content-type': 'application/json'
       },
-      body: body,
-      json: true
+      json: body
     };
 
-    let self = this;
-    await httppromise(options).then(function (response: any) {
-
-      self.stitchdata.AccountId = response.stitch_account_id;
-      self.stitchdata.Token = response.access_token;
+    try {
+      const response: any = await httppromise(options);
+      this.stitchdata.AccountId = response.stitch_account_id;
+      this.stitchdata.Token = response.access_token;
       Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.Stitchdata, Constants.StitchdataCreateAccountSuccess, response, body));
-    }).catch(function (err: any) {
+      return new StitchdataAccountAuthorizationSuccessResponse(this.stitchdata);
+    } catch (err) {
       Logger.updateLogs(new Log(EnumCurrentStatus.Error, EnumModule.Stitchdata, Constants.StitchdataCreateAccountError, err, body));
-    })
+    }
+    
+  }
 
-
-    return new StitchdataAccountAuthorizationSuccessResponse(this.stitchdata);
-  };
 }
 
 
