@@ -1,7 +1,7 @@
 import { Stitchdata } from '../class/Stitchdata';
 import { BaseLayer } from './BaseLayer';
-const httppromise = require('request-promise');
-var sessionstorage = require('sessionstorage');
+import httppromise, { Options } from 'got';
+import sessionstorage from 'sessionstorage';
 import { Log } from '../class/Log'
 import { Logger } from '../class/Logger'
 import { EnumCurrentStatus } from '../Enum/EnumCurrentStatus'
@@ -9,15 +9,13 @@ import Constants from '../class/Constants'
 import { EnumModule } from '../Enum/EnumModule';
 import { EnumToken } from '../Enum/EnumToken';
 
-
-
 export class StitchdataBaseLayer extends BaseLayer {
 
     constructor() {
         super();
     }
 
-    baseUrl(url: string) {
+    baseUrl(url: string): string {
         return this.environmentConfig.Stitchdata.Urls.BaseUrl + url;
     }
 
@@ -30,7 +28,7 @@ export class StitchdataBaseLayer extends BaseLayer {
         this.stitchdata.GrantType = this.environmentConfig.Stitchdata.GrantType;
         this.stitchdata.ContentType = this.environmentConfig.Stitchdata.ContentType;
 
-        var options = {
+        const options: Options = {
             method: 'POST',
             url: this.baseUrl(Constants.StitchdataAuthURL),
             headers:
@@ -46,44 +44,40 @@ export class StitchdataBaseLayer extends BaseLayer {
         };
 
         return httppromise(options);
-    };
+    }
 
     //Check if Stitchdata token already exists otherwise it will generate new Token for Stitchdata
-    protected async authorizeStitchdata(stitchdata: Stitchdata) {
-
+    protected async authorizeStitchdata(stitchdata: Stitchdata): Promise<boolean> {
         if (sessionstorage.getItem(EnumToken.StitchdataToken) == null) {
-
-            return this.generateStitchdataToken()
-                .then(function (response: any) {
-                    let jsonResponse = JSON.parse(response);
-                    stitchdata.AccountId = jsonResponse.stitch_account_id;
-                    stitchdata.Token = jsonResponse.access_token;
-                    stitchdata.TokenType = jsonResponse.token_type;
-                    //temporary commented sessionstorage.setItem(EnumToken.StitchdataToken, jsonResponse.access_token)
-                    Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.Stitchdata, Constants.StitchdataAuthSuccess, response, ""))
-                    return true;
-                })
-                .catch(function (err: any) {
-                    Logger.updateLogs(new Log(EnumCurrentStatus.Error, EnumModule.Stitchdata, Constants.StitchdataAuthError, err, ""));
-                    return false;
-                });
+            try {
+                const response: any = await this.generateStitchdataToken();
+                const jsonResponse = JSON.parse(response);
+                stitchdata.AccountId = jsonResponse.stitch_account_id;
+                stitchdata.Token = jsonResponse.access_token;
+                stitchdata.TokenType = jsonResponse.token_type;
+                //temporary commented sessionstorage.setItem(EnumToken.StitchdataToken, jsonResponse.access_token)
+                Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.Stitchdata, Constants.StitchdataAuthSuccess, response, ""))
+                return true;
+            } catch (err) {
+                Logger.updateLogs(new Log(EnumCurrentStatus.Error, EnumModule.Stitchdata, Constants.StitchdataAuthError, err, ""));
+            }
         }
+        return false;
     }
 
 
     //Setup stitch object essentials
-    protected assignStitchdataEssentials() {
+    protected assignStitchdataEssentials(): void {
         this.stitchdata.PartnerId = this.environmentConfig.Stitchdata.PartnerId;
         this.stitchdata.PartnerSecret = this.environmentConfig.Stitchdata.PartnerSecret;
     }
 
     //remove token
-    protected removeToken() {
+    protected removeToken(): void {
         try {
             sessionstorage.removeItem(EnumToken.StitchdataToken);
             Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.Stitchdata, Constants.StitchdataTokenRemovedSuccess, "", ""))
-        }
-        catch (error) {
+        } catch (error) {
             console.error(error);
             Logger.updateLogs(new Log(EnumCurrentStatus.Error, EnumModule.Stitchdata, Constants.StitchdataTokenRemovedError, "", ""));
         }
