@@ -1,5 +1,4 @@
-const httppromise = require('request-promise');
-var sessionstorage = require('sessionstorage');
+import httppromise, { Options } from 'got';
 import { Log } from '../class/Log'
 import { Logger } from '../class/Logger'
 import { EnumCurrentStatus } from '../Enum/EnumCurrentStatus'
@@ -18,33 +17,33 @@ class StitchdataHttpRequests extends StitchdataBaseLayer {
   }
 
   //Here we are generating token for existing account
-  async generateTokenForExistingStitchdataAccount(RequestBody: any) {
+  async generateTokenForExistingStitchdataAccount(requestBody: any) {
 
     this.removeToken();
-    this.stitchdata.AuthorizationCode = RequestBody.stitch_auth_code;
-    let self = this;
-    await this.authorizeStitchdata(self.stitchdata).then(function (response: any) {
-      Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.Stitchdata, Constants.StitchdataAuthError, response, ""));
-    }).catch(function (err: any) {
+    this.stitchdata.AuthorizationCode = requestBody.stitch_auth_code;
+    try {
+      const response: any = await this.authorizeStitchdata(this.stitchdata);
+      Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.Stitchdata, Constants.StitchdataAuthError, response as any, ""));
+      return new StitchdataAccountAuthorizationSuccessResponse(this.stitchdata);
+    } catch (err) {
       Logger.updateLogs(new Log(EnumCurrentStatus.Error, EnumModule.Stitchdata, Constants.StitchdataAuthSuccess, err, ""));
-    });
+    }
 
-    return new StitchdataAccountAuthorizationSuccessResponse(this.stitchdata);
-  };
+  }
 
 
   //Here we are creating new Account
-  async registerAccountForStitchdata(RequestBody: any) {
+  async registerAccountForStitchdata(requestBody: any) {
 
     this.assignStitchdataEssentials();
-    let stitchdataCreateAccount: StitchdataCreateAccount = new StitchdataCreateAccount();
+    const stitchdataCreateAccount = new StitchdataCreateAccount();
 
-    stitchdataCreateAccount.FirstName = RequestBody.firstname;
-    stitchdataCreateAccount.LastName = RequestBody.lastname;
-    stitchdataCreateAccount.Email = RequestBody.email;
-    stitchdataCreateAccount.Company = RequestBody.company;
+    stitchdataCreateAccount.FirstName = requestBody.firstname;
+    stitchdataCreateAccount.LastName = requestBody.lastname;
+    stitchdataCreateAccount.Email = requestBody.email;
+    stitchdataCreateAccount.Company = requestBody.company;
 
-    let body = {
+    const body = {
       "partner_id": this.stitchdata.PartnerId,
       "partner_secret": this.stitchdata.PartnerSecret,
       "first_name": stitchdataCreateAccount.FirstName,
@@ -53,84 +52,70 @@ class StitchdataHttpRequests extends StitchdataBaseLayer {
       "email": stitchdataCreateAccount.Email,
     }
 
-    let options = {
+    const options: Options = {
       url: this.baseUrl(Constants.StitchdataCreateAccountURL),
       method: 'POST',
       headers: {
         'content-type': 'application/json'
       },
-      body: body,
-      json: true
+      json: body
     };
 
-    let self = this;
-
-    await httppromise(options).then(function (response: any) {
-
-      self.stitchdata.AccountId = response.stitch_account_id;
-      self.stitchdata.Token = response.access_token;
+    try {
+      const response: any = await httppromise(options)
+      this.stitchdata.AccountId = response.stitch_account_id;
+      this.stitchdata.Token = response.access_token;
       Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.Stitchdata, Constants.StitchdataCreateAccountSuccess, response, body));
-    }).catch(function (err: any) {
-
+      return new StitchdataAccountCreationSuccessResponse(this.stitchdata, stitchdataCreateAccount);
+    } catch (err) {
       Logger.updateLogs(new Log(EnumCurrentStatus.Error, EnumModule.Stitchdata, Constants.StitchdataCreateAccountError, err, body));
-    })
+    }
 
-    return new StitchdataAccountCreationSuccessResponse(this.stitchdata, stitchdataCreateAccount);
-  };
-
+  }
 
   //Here we are retrieving all the sources from new Account
-  async retrievesourcesfromstitchdata(RequestBody: any) {
+  async retrievesourcesfromstitchdata(requestBody: any) {
 
-    let options = {
+    const options: Options = {
       url: this.baseUrl(Constants.StitchdataRetrieveSourcesURL),
       method: 'GET',
       headers: {
-        'Authorization': RequestBody.access_token,
+        'Authorization': requestBody.access_token,
         'content-type': 'application/json'
-      },
-      json: true
+      }
     };
-
-    let responseStitchdata;
-    await httppromise(options).then(function (response: any) {
-
-      responseStitchdata = response;
+    try {
+      const response: any = await httppromise(options)
       Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.Stitchdata, Constants.StitchdataRetrieveSourcesSuccess, response, options));
-    }).catch(function (err: any) {
+      return new StitchdataGetRequestSuccessResponse(requestBody.access_token, response);
+    } catch (err) {
       Logger.updateLogs(new Log(EnumCurrentStatus.Error, EnumModule.Stitchdata, Constants.StitchdataRetrieveSourcesError, err, options));
-    })
+    }
 
-
-    return new StitchdataGetRequestSuccessResponse(RequestBody.access_token, responseStitchdata);
-  };
-
+  }
 
   //Here we are retrieving all the sources from new Account
-  async retrievedestinationfromstitchdata(RequestBody: any) {
+  async retrievedestinationfromstitchdata(requestBody: any) {
 
-    let options = {
+    const options: Options = {
       url: this.baseUrl(Constants.StitchdataRetrieveDestinationsURL),
       method: 'GET',
       headers: {
-        'Authorization': RequestBody.access_token,
+        'Authorization': requestBody.access_token,
         'content-type': 'application/json'
-      },
-      json: true
+      }
     };
 
-    let responseStitchdata;
-    await httppromise(options).then(function (response: any) {
-
-      responseStitchdata = response;
+    try {
+      const response: any = await httppromise(options);
       Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.Stitchdata, Constants.StitchdataRetrieveSourcesSuccess, response, options));
-    }).catch(function (err: any) {
+      return new StitchdataGetRequestSuccessResponse(requestBody.access_token, response);
+    } catch (err) {
       Logger.updateLogs(new Log(EnumCurrentStatus.Error, EnumModule.Stitchdata, Constants.StitchdataRetrieveSourcesError, err, options));
-    })
+    }
 
-    return new StitchdataGetRequestSuccessResponse(RequestBody.access_token, responseStitchdata);
-  };
+  }
+
 }
-
 
 export default new StitchdataHttpRequests();
