@@ -1,7 +1,7 @@
 import { Veeam } from '../class/Veeam';
 import { BaseLayer } from './BaseLayer';
-const httppromise = require('request-promise');
-var sessionstorage = require('sessionstorage');
+import httppromise, { Options } from 'got';
+import sessionstorage from 'sessionstorage';
 import { Log } from '../class/Log'
 import { Logger } from '../class/Logger'
 import { EnumCurrentStatus } from '../Enum/EnumCurrentStatus'
@@ -12,12 +12,13 @@ import { EnumToken } from '../Enum/EnumToken';
 export class VeeamBaseLayer extends BaseLayer {
 
     protected veeam: Veeam = new Veeam();
+
     constructor() {
         super();
         this.setDefaultValues();
     }
 
-    setDefaultValues() {
+    setDefaultValues(): void {
         //below valaues are default and pickedup from config file
         this.veeam.CloudConnectAgentUid = this.environmentConfig.Veeam.CloudConnectAgentUid;
         this.veeam.BackupProtectionEnabled = this.environmentConfig.Veeam.DefaultValues.BackupProtectionEnabled;
@@ -26,42 +27,42 @@ export class VeeamBaseLayer extends BaseLayer {
         this.veeam.ExpirationEnabled = this.environmentConfig.Veeam.DefaultValues.ExpirationEnabled;
     }
 
-    baseUrl(url: string) {
+    baseUrl(url: string): string {
         return this.environmentConfig.Veeam.Urls.BaseUrl + url;
     }
 
     //Generate new Token for veeam
     private generateVeeamToken() {
         //let body = "grant_type=password&username=TESTING.ASEIT.NET%5Cmunish.singla&password=munishtest2019";
-        let body = "grant_type=password&username=" + this.environmentConfig.Veeam.Username + "&password=" + this.environmentConfig.Veeam.Password;
+        const body = "grant_type=password&username=" + this.environmentConfig.Veeam.Username + "&password=" + this.environmentConfig.Veeam.Password;
 
-        let options = {
+        const options: Options = {
             url: this.baseUrl(Constants.VeeamAuthURL),
             method: 'POST',
             body: body
         };
         return httppromise(options);
-    };
+    }
 
     //Check if Veeam token already exists otherwise it will generate new Token for Veeam
-    protected authorizeVeeamGrid() {
-        
+    protected async authorizeVeeamGrid(): Promise<boolean> {
         if (sessionstorage.getItem(EnumToken.VeeamToken) == null) {
-                return this.generateVeeamToken().then(function (response: any) {
-                let veemAuthenticationResponse = JSON.parse(response);
+            try {
+                const response: any = await this.generateVeeamToken();
+                const veemAuthenticationResponse = JSON.parse(response);
                 sessionstorage.setItem(EnumToken.VeeamToken, "Bearer " + veemAuthenticationResponse.access_token)
                 Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.Veeam, Constants.VeeamAuthSuccess, response, ""))
                 return true;
-            }).catch(function (err: any) {
+            } catch (err) {
                 Logger.updateLogs(new Log(EnumCurrentStatus.Error, EnumModule.Veeam, Constants.VeeamAuthError, err, ""));
-                return false;
-            })
+            }
         }
+        return false;
     }
 
     //isAuthorized token
-    protected async isAuthorized() {
-         if (await Logger.hasErrorLogs() == true)
+    protected async isAuthorized(): Promise<boolean> {
+        if (await Logger.hasErrorLogs() == true)
             return false;
         else if (await this.authorizeVeeamGrid() == false)
             return false;
@@ -71,7 +72,7 @@ export class VeeamBaseLayer extends BaseLayer {
 
 
     //remove token
-    protected removeToken() {
+    protected removeToken(): void {
         try {
             sessionstorage.removeItem(EnumToken.VeeamToken);
             Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.Veeam, Constants.VeeamTokenRemovedSuccess, "", ""))
