@@ -1,7 +1,6 @@
 import { BaseLayer } from './BaseLayer';
-const request = require('request');
-const httppromise = require('request-promise');
-var sessionstorage = require('sessionstorage');
+import httppromise, { Options } from 'got';
+import sessionstorage from 'sessionstorage';
 import { Log } from '../class/Log'
 import { Logger } from '../class/Logger'
 import { EnumCurrentStatus } from '../Enum/EnumCurrentStatus'
@@ -15,15 +14,15 @@ export class XcloudBaseLayer extends BaseLayer {
         super();
     }
 
-    baseUrl(url: string) {
+    baseUrl(url: string): string {
         return this.environmentConfig.Xcloud.Urls.BaseUrl + url;
     }
 
     //Generate new Token for Xcloud
-    private generateXcloudToken() {
-        let body = "user=" + this.environmentConfig.Xcloud.Username + "&password=" + this.environmentConfig.Xcloud.Password + "&auth_scheme_id=" + this.environmentConfig.Xcloud.auth_scheme_id;
+    private generateXcloudToken(): Options {
+        const body = "user=" + this.environmentConfig.Xcloud.Username + "&password=" + this.environmentConfig.Xcloud.Password + "&auth_scheme_id=" + this.environmentConfig.Xcloud.auth_scheme_id;
 
-        let options = {
+        return {
             url: this.baseUrl(Constants.XcloudAuthURL),
             method: 'POST',
             'headers': {
@@ -31,32 +30,26 @@ export class XcloudBaseLayer extends BaseLayer {
             },
             body: body
         };
-
-        return options;
-    };
+    }
 
     //Check if Xcloud token already exists otherwise it will generate new Token for Xcloud
-    protected async authorizeXcloudGrid() {
+    protected async authorizeXcloudGrid(): Promise<any> {
 
-        let options = this.generateXcloudToken();
-        return new Promise((resolve, reject) => {
-            request(options, function (error: any, response: any) {
-                if (error) {
-                    console.log('Error')
-                    throw new Error(error);
-                }
-                else {
-                    let cookie = ((response.headers["set-cookie"]).toString()).split(";")[0];
-                    sessionstorage.setItem(EnumToken.XcloudCookie, cookie)
-                    return resolve({ connect_sid: cookie })
-                }
-            });
-        })
+        const options = this.generateXcloudToken();
+        try {
+            const response: any = await httppromise(options);
+            const cookie = ((response.headers["set-cookie"]).toString()).split(";")[0];
+            sessionstorage.setItem(EnumToken.XcloudCookie, cookie);
+            return { connect_sid: cookie };
+        } catch (err) {
+            console.log('Error')
+            throw new Error(err);
+        }
 
     }
 
     //isAuthorized token
-    protected async isAuthorized() {
+    protected async isAuthorized(): Promise<boolean> {
         sessionstorage.removeItem(EnumToken.XcloudCookie);
         if (await Logger.hasErrorLogs() == true)
             return false;
@@ -67,7 +60,7 @@ export class XcloudBaseLayer extends BaseLayer {
     }
 
     //remove token
-    protected removeToken() {
+    protected removeToken(): void {
         try {
             sessionstorage.removeItem(EnumToken.XcloudCookie);
             Logger.updateLogs(new Log(EnumCurrentStatus.Success, EnumModule.Xcloud, Constants.XcloudTokenRemovedSuccess, "", ""))
