@@ -1,3 +1,3 @@
 export enum EnumCurrency {
   AUD = "1"
-  }
+}
