@@ -1,4 +1,4 @@
 export enum EnumCurrentStatus {
   Error = "Error",
   Success = "Success"
-  }
+}
