@@ -1,3 +1,3 @@
 export enum EnumCustomForm {
   ASEClientForm = "4"
-  }
+}
