@@ -1,12 +1,5 @@
 import { EnumModule } from '../Enum/EnumModule'
 
 export class BaseResponse {
-    module: EnumModule;
-   // message: string;
-
-    constructor(){
-        this.module = EnumModule.Storagegrid;
-        //this.message = "";
-    }
+    module: EnumModule = EnumModule.Storagegrid;
 }
-
