@@ -1,10 +1,5 @@
 export class Bucket {
-    name: string;
-    region: string;
-
-    constructor() {
-        this.name = "";
-        this.region = "";
-    }
+    name = "";
+    region = "";
 }
 
