@@ -3,15 +3,13 @@ import { EnumModule } from '../../Enum/EnumModule'
 import { BaseResponse } from '../BaseResponse'
 
 export class DataikuListDatasetsSuccessResponse extends BaseResponse {
-    
-    response: any[];
+
+    response: unknown[];
 
     constructor(dataikuDatasets: any) {
         super();
         this.module = EnumModule.Dataiku;
         this.response = dataikuDatasets;
-        
-        return this;
     }
 }
 
