@@ -6,14 +6,13 @@ export class StitchdataAccountAuthorizationSuccessResponse extends BaseResponse 
     access_token: string;
     token_type: string;
     stitch_account_id: string;
-    
+
     constructor(stitchdata: Stitchdata) {
         super();
         this.module = EnumModule.Stitchdata;
         this.access_token = stitchdata.Token;
         this.token_type = stitchdata.TokenType;
         this.stitch_account_id = stitchdata.AccountId;
-        return this;
     }
 }
 
