@@ -1,18 +1,14 @@
 import { BaseResponse } from "./BaseResponse";
 import { EnumModule } from "../Enum/EnumModule";
 
-export class StitchdataCreateAccount extends BaseResponse  {
-    FirstName: string;
-    LastName: string;
-    Company: string;
-    Email: string;
+export class StitchdataCreateAccount extends BaseResponse {
+    FirstName = "";
+    LastName = "";
+    Company = "";
+    Email = "";
     constructor() {
         super();
         this.module = EnumModule.Stitchdata;
-        this.FirstName = "";
-        this.LastName = "";
-        this.Company = ""
-        this.Email = ""
     }
 }
 
