@@ -1,11 +1,6 @@
 export class TenantPolicy {
-    UseAccountIdentitySource: string;
-    AllowPlatformServices: string;
-    QuotaObjectBytes: string;
-    constructor() {
-        this.UseAccountIdentitySource = "";
-        this.AllowPlatformServices = "";
-        this.QuotaObjectBytes = "";
-    }
+    UseAccountIdentitySource = "";
+    AllowPlatformServices = "";
+    QuotaObjectBytes = "";
 }
 
